pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quickview"
version = "0.1.0"
description = "Terminal file viewer for CSV, Excel, PDF, Word, ZIP, audio, and images"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Sebastian Stianek" },
]
keywords = [
    "terminal",
    "file-viewer",
    "csv",
    "excel",
    "pdf",
    "audio",
    "waveform",
    "tui",
    "cli",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

dependencies = [
    "textual>=0.40.0",
]

[project.optional-dependencies]
excel = ["openpyxl>=3.0.0"]
pdf = ["pypdf>=3.0.0"]
image = ["Pillow>=9.0.0"]
docx = ["python-docx>=1.0.0"]
audio = ["pydub>=0.25.0"]

# Install all optional dependencies
all = [
    "openpyxl>=3.0.0",
    "pypdf>=3.0.0",
    "Pillow>=9.0.0",
    "python-docx>=1.0.0",
    "pydub>=0.25.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
quickview = "quickview.cli:main"
qv = "quickview.cli:main"

[project.urls]
Homepage = "https://github.com/sebsastianek/quickview"
Repository = "https://github.com/sebsastianek/quickview"
Issues = "https://github.com/sebsastianek/quickview/issues"

[tool.hatch.build.targets.wheel]
packages = ["quickview"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
